Email configuration method change
Old email configuration method based in context resources changed to a
new method that configures the Java Mail Session using the
getDefault(Properties) method, where properties are taken from the
web.xml file.

The reason for this change is that with this new method the hosting
Tomcat doesn't need the Java Mail libraries on its path.